+++ lib/utils/server/feature-flags.ts
@@ -1,7 +1,7 @@
 export type FeatureFlag = "contributions_evolution_by_type" | "workspaces";
 
 export async function getAllFeatureFlags(userId: number) {
-  const response = await fetch("https://app.posthog.com/decide?v=3", {
+  const requestOptions = {
     method: "POST",
     headers: {
       "Content-Type": "application/json",
@@ -10,8 +10,12 @@ export async function getAllFeatureFlags(userId: number) {
       api_key: process.env.NEXT_PUBLIC_POSTHOG_ID,
       distinct_id: userId,
     }),
-  });
+  };
 
+  const response = await fetch(
+    "https://app.posthog.com/decide?v=3", // or eu
+    requestOptions
+  );
   const { featureFlags } = await response.json();
 
   return featureFlags;